+++ cypress/e2e/form-level/form-level.cy.js
@@ -52,7 +52,7 @@ describe('Home page', () => {
     it('should import LOINC form', () => {
       cy.get('input[type="radio"][value="loinc"]').should('be.visible').click();
       cy.contains('button', 'Continue').click();
-      cy.get('#loincSearch').type('vital');
+      cy.get('#loincSearch').type('vital signs with');
       cy.get('ngb-typeahead-window').should('be.visible');
       cy.get('ngb-typeahead-window button').first().click();
       cy.get('#title').should('have.value', 'Vital signs with method details panel');
@@ -131,7 +131,7 @@ describe('Home page', () => {
       cy.contains('.mat-tab-label-content', 'View Rendered Form').scrollIntoView().click();
       cy.get('wc-lhc-form').should('exist', true, {timeout: 10000});
       cy.get('#\\/54126-8\\/54133-4\\/1\\/1').as('ethnicity');
-      cy.get('@ethnicity').scrollIntoView().type('latin');
+      cy.get('@ethnicity').scrollIntoView().type('lat');
       cy.get('#completionOptions').should('be.visible', true);
       cy.get('@ethnicity').type('{downarrow}');
       cy.get('@ethnicity').type('{enter}');
@@ -194,7 +194,7 @@ describe('Home page', () => {
       cy.get('@add').should('have.attr', 'disabled');
 
       cy.get('@inputUrl').clear();
-      cy.get('@inputUrl').type('http://localhost'); // Valid format, but a FHIR server.
+      cy.get('@inputUrl').type('http://localhost'); // Valid format, but not a FHIR server.
       cy.get('@validate').click();
       cy.contains('p.text-danger', 'Unable to confirm that that URL is a FHIR server.').should('be.visible', true);
       cy.get('@add').should('have.attr', 'disabled');