+++ src/app/services/fetch.service.ts
@@ -5,8 +5,6 @@ import { Injectable } from '@angular/core';
 import {HttpClient, HttpParams, HttpResponse} from '@angular/common/http';
 import {Observable, of} from 'rxjs';
 import {catchError, map, tap} from 'rxjs/operators';
-import {TreeNode, ITreeOptions} from '@circlon/angular-tree-component/';
-import {TREE_ACTIONS, KEYS, TreeModel} from '@circlon/angular-tree-component';
 import {AutoCompleteResult} from '../lib/widgets/auto-complete/auto-complete.component';
 import {Util} from '../lib/util';
 import {fhir} from '../fhir';