+++ src/app/testing/common-testing.module.ts
@@ -1,6 +1,6 @@
 import {ChangeDetectorRef, Component, Injectable, Input, NgModule, OnInit} from '@angular/core';
 import {ComponentFixture, TestBed, waitForAsync} from '@angular/core/testing';
-import {TreeModule} from '@circlon/angular-tree-component';
+import {TreeModule} from '@bugsplat/angular-tree-component';
 import {
   DefaultWidgetRegistry, FormProperty,
   FormPropertyFactory,